% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlp_allocate.R
\name{AllocateMon}
\alias{AllocateMon}
\title{Monthly C and/or N allocation for the PnET model}
\usage{
AllocateMon(sitepar, vegpar, share, rstep, model = "pnet-ii")
}
\arguments{
\item{sitepar}{A table that contains site-specific variables.}

\item{vegpar}{A table that contains vegetation-specific variables.}

\item{share}{The shared object containing intermittent variables.}

\item{rstep}{current time step}
}
\description{
The following variables are calculated/updated:
\itemize{
\item RootC
\item PlantC
\item WoodMRespYr
\item FolProdCYr
\item FolGRespYr
\item GDDWoodEff
\item WoodProdCYr
\item WoodGRespYr
\item RootProdCYr
\item RootMRespYr
\item RootGRespYr
\item NetCBal
--- For PnET-CN ---
\itemize{
\item WoodMass
\item WoodMassN
\item RootMass
\item RootMassN
\item PlantN
}
}
}
