% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlp_allocate.R
\name{AllocateYr}
\alias{AllocateYr}
\title{Annual C and/or N allocation}
\usage{
AllocateYr(sitepar, vegpar, share, rstep, model = "pnet-ii")
}
\arguments{
\item{sitepar}{A table that contains site-specific variables.}

\item{vegpar}{A table that contains vegetation-specific variables.}

\item{share}{The shared object containing intermittent variables.}

\item{rstep}{current time step}
}
\description{
The following variables are calculated/updated:
\itemize{
\item BudC
\item WoodC
\item PlantC
\item NPPFolYr
\item NPPWoodYr
\item NPPRootYr
\item FolMassMax (in VegPar)
\item FolMassMin (in VegPar)
\item NEP
--- For PnET-CN only ---
\itemize{
\item PlantN
\item FolNCon (in VegPar)
\item NH4
\item NRatio
\item RootNSinkEff
\item FolN
\item FolC
\item TotalN
\item TotalM
}
}
}
