% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlp_atm_environ.R
\name{AtmEnviron}
\alias{AtmEnviron}
\title{Calculate atmospheric environmental variables}
\usage{
AtmEnviron(climate_dt, lat, share_dt)
}
\arguments{
\item{climate_dt}{Climate data table.}

\item{lat}{Site latitude.}

\item{share_dt}{The shared data table from a \code{share} object used to store
intermittend variables.}
}
\value{
A data.table w/ additional columns compared to the \code{climate_dt}.
}
\description{
The following variables are calculated/updated:
\itemize{
\item Tavg
\item Tday
\item Tnight
\item VPD
\item Tmin
\item Month
\item Dayspan
\item Daylenhr
\item Daylen
\item Nightlen
\item GDD
\item GDDTot
}
}
