% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{CalRealizedResp}
\alias{CalRealizedResp}
\title{Calculate reslized respiration per unit mass.}
\usage{
CalRealizedResp(BaseFolResp, RespQ10, T, PsnTOpt, TimeLen)
}
\arguments{
\item{BaseFolResp}{Base foliar respiration rate.}

\item{RespQ10}{The Q10 parameter.}

\item{T}{Temperature.}

\item{PsnTOpt}{Optimum temperature for photosynthesis.}

\item{TimeLen}{The length of time, e.g, day length or night length.}
}
\value{
Realized respiration per unit mass.
}
\description{
Calculate reslized respiration per unit mass.
}
