% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlp_cntrans.R
\name{CalRootTurnoverRate}
\alias{CalRootTurnoverRate}
\title{Calculate root turnover rate (Aber et al 1997).}
\usage{
CalRootTurnoverRate(vegpar, netNmin, dayspan, NumYrDays, BiomLossFrac)
}
\arguments{
\item{vegpar}{The vegetation parameter object.}

\item{netNmin}{Last year's N mineralization.}

\item{dayspan}{Current month's number of days.}

\item{BiomLossFrac}{Biome loss fraction due to disturbance}
}
\value{
Root turnover rate.
}
\description{
Calculate root turnover rate (Aber et al 1997).
}
