% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlp_cntrans.R
\name{CNTrans}
\alias{CNTrans}
\title{CNTrans routine in PnET-CN}
\usage{
CNTrans(climate_dt, sitepar, vegpar, share, rstep)
}
\arguments{
\item{climate_dt}{A table that contains monthly climate data.}

\item{sitepar}{A table that contains site-specific variables.}

\item{vegpar}{A table that contains vegetation-specific variables.}

\item{share}{The shared object containing intermittent variables.}

\item{rstep}{current time step}
}
\description{
The following variables are calculated/updated:
\itemize{
\item HOM
\item HON
\item RootMass
\item RootMassN
\item WoodMass
\item WoodMassN
\item WoodDecResp
\item WoodDecRespYr
\item DeadWoodM
\item DeadWoodN
\item NetCBal
\item PlantN
\item FolLitM
\item FolMass
\item PlantC
\item TotalLitterM
\item TotalLitterMYr
\item TotalLitterN
\item TotalLitterNYr
}
}
