% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlp_sens_analysis.R
\name{DoSens}
\alias{DoSens}
\title{This is an integreated function to perfom sensitivity analysis.}
\usage{
DoSens(
  clim_dt,
  sitepar,
  vegpar,
  model,
  outfile,
  target_pars = c("CanopyGrossPsn", "CanopyNetPsn"),
  annual = TRUE
)
}
\arguments{
\item{clim_dt}{A table that contains monthly climate data.}

\item{sitepar}{A table that contains site-specific variables.}

\item{vegpar}{A table that contains vegetation-specific variables.}

\item{model}{Character. Case insensitive. One of "pnet-day", "pnet-ii",
and "pnet-cn".}

\item{outfile}{A path indicating a pdf file for plotting out the
parameter importance figures.}

\item{target_pars}{Interested target output parameters.}

\item{annual}{Logical. Whether the target parameters are annual.}
}
\description{
This is an integreated function to perfom sensitivity analysis.
}
