% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlp_leach.R
\name{Leach}
\alias{Leach}
\title{The Leach module in PnET-CN}
\usage{
Leach(share, rstep)
}
\arguments{
\item{share}{The shared object containing intermittent variables.}

\item{rstep}{current time step}
}
\description{
The following variables are calculated/updated:
\itemize{
\item NDrainYr
\item NO3
}
}
