% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlp_pheno.R
\name{Phenology}
\alias{Phenology}
\title{Phenology module}
\usage{
Phenology(sitepar, vegpar, share, rstep, phenophase)
}
\arguments{
\item{sitepar}{A table that contains site-specific variables.}

\item{vegpar}{A table that contains vegetation-specific variables.}

\item{share}{The shared object containing intermittent variables.}

\item{rstep}{current time step}

\item{phenophase}{Phenology stage, can be "grow" or "senesce".}
}
\description{
The following variables are calculated/updated:
\itemize{
\item GDDFolEff
\item FolMass
\item FolProdCMo
\item FolGRespMo
\item PosCBalMass
\item LAI
\item FolLitM
}
}
