% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlp_psn.R
\name{Photosynthesis}
\alias{Photosynthesis}
\title{The Photosynthesis module}
\usage{
Photosynthesis(climate_dt, sitepar, vegpar, share, rstep, model = "pnet-ii")
}
\arguments{
\item{climate_dt}{A table that contains monthly climate data.}

\item{sitepar}{A table that contains site-specific variables.}

\item{vegpar}{A table that contains vegetation-specific variables.}

\item{share}{The shared object containing intermittent variables.}

\item{rstep}{current time step}
}
\description{
The following variables are calculated/updated:
\itemize{
\item DTemp
\item GrossAmax
\item PosCBalMass
\item LAI
\item CanopyNetPsn
\item CanopyGrossPsn
\item PosCBalMassTot
\item PosCBalMassIx
\item LightEffMin (in Share)
}
}
