% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pnet_cn.R
\name{PnET_CN}
\alias{PnET_CN}
\title{The PnET-CN model}
\usage{
PnET_CN(climate_dt, sitepar, vegpar, verbose = FALSE)
}
\arguments{
\item{climate_dt}{A table that contains monthly climate data.}

\item{sitepar}{A table that contains site-specific variables.}

\item{vegpar}{A table that contains vegetation-specific variables.}
}
\value{
A list that contains both annual and monthly computed variables.
}
\description{
Please see \href{https://linkinghub.elsevier.com/retrieve/pii/S0304380097019534}{Aber et al 1997} for
context of the model. And, \code{doc/pnet_model.md} file contains technical
details of the model.
}
