% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pnet_day.R
\name{PnET_Day}
\alias{PnET_Day}
\title{The PnET-Day model}
\usage{
PnET_Day(climate_dt, sitepar, vegpar, verbose = FALSE)
}
\arguments{
\item{climate_dt}{A table that contains monthly climate data.}

\item{sitepar}{A table that contains site-specific variables.}

\item{vegpar}{A table that contains vegetation-specific variables.}
}
\value{
A list that contains both annual and monthly computed variables.
}
\description{
Please see \href{http://www.jstor.org/stable/4221255}{Aber et al 1996}
for context of the model. And, \code{doc/pnet_model.md} file contains technical
details of the model. Note that PnET-Day does not have water balance and
allocation, it only simulates photosynthesis without water stress.
}
