% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pnet_ii.R
\name{PnET_II}
\alias{PnET_II}
\title{The PnET-II model}
\usage{
PnET_II(climate_dt, sitepar, vegpar, verbose = FALSE)
}
\arguments{
\item{climate_dt}{A table that contains monthly climate data.}

\item{sitepar}{A table that contains site-specific variables.}

\item{vegpar}{A table that contains vegetation-specific variables.}
}
\value{
A list that contains both annual and monthly computed variables.
}
\description{
Please see \href{https://doi.org/10.1007/BF00317837}{Aber et al 1992}
and \href{http://www.int-res.com/abstracts/cr/v05/n3/p207-222/}{Aber et al 1995}
for context of the model. And, \code{doc/pnet_ii_model.md} file contains technical
details of the model.
}
