% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_site_par.R
\name{SitePar}
\alias{SitePar}
\title{The site related parameters.}
\description{
These default values are from Aber et al 1992, 1995, 1996
}
\section{Super class}{
\code{pnetr::Param} -> \code{SitePar}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SitePar-clone}{\code{SitePar$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="pnetr" data-topic="Param" data-id="initialize"><a href='../../pnetr/html/Param.html#method-Param-initialize'><code>pnetr::Param$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SitePar-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SitePar-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SitePar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
