% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlp_soil_resp.R
\name{SoilRespiration}
\alias{SoilRespiration}
\title{Soil respiration module}
\usage{
SoilRespiration(sitepar, vegpar, share, rstep)
}
\arguments{
\item{sitepar}{A table that contains site-specific variables.}

\item{vegpar}{A table that contains vegetation-specific variables.}

\item{share}{The shared object containing intermittent variables.}

\item{rstep}{current time step}

\item{phenophase}{Phenology stage, can be "grow" or "senesce".}
}
\description{
The following variables are calculated/updated:
\itemize{
\item SoilRespMo
\item SoilRespYr
}
}
