% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_veg_par.R
\name{VegPar}
\alias{VegPar}
\title{The vegetation related parameters.}
\description{
The vegetation related parameters.

The vegetation related parameters.
}
\section{Super class}{
\code{pnetr::Param} -> \code{VegPar}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VegPar-clone}{\code{VegPar$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="pnetr" data-topic="Param" data-id="initialize"><a href='../../pnetr/html/Param.html#method-Param-initialize'><code>pnetr::Param$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VegPar-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VegPar-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VegPar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
