% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlp_sens_analysis.R
\name{VisSense}
\alias{VisSense}
\title{Visualize the sensitivity analysis.}
\usage{
VisSense(par_mat, sens_res_dt, outfile)
}
\arguments{
\item{par_mat}{The input parameter matrix for the sensitivity analysis.}

\item{sens_res_dt}{The output sensitivity analysis data.table.}

\item{outfile}{A pdf filename to visualize the figures.}
}
\description{
This function uses a random forest model to fit the model
outputs with input parameters to determine the importance of the input
parameters.
}
