% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlp_water_bal.R
\name{Waterbal}
\alias{Waterbal}
\title{The Water Balance module}
\usage{
Waterbal(climate_dt, sitepar, vegpar, share, rstep, model = "pnet-ii")
}
\arguments{
\item{climate_dt}{A table that contains monthly climate data.}

\item{sitepar}{A table that contains site-specific variables.}

\item{vegpar}{A table that contains vegetation-specific variables.}

\item{share}{The shared object containing intermittent variables.}

\item{rstep}{current time step}
}
\description{
The following variables are calculated/updated:
\itemize{
\item SnowPack (in SitePar)
\item Water
\item DWater
\item DWatertot
\item DWaterIx
\item MeanSoilMoistEff
\item CanopyGrossPsnActMo
\item GrsPsnMo
\item NetPsnMo
\item Drainage
\item TotTrans
\item TotPsn
\item TotDrain
\item TotPrec
\item TotEvap
\item TotGrossPsn
\item ET
---- For PnET-CN
\item CanopyDO3Pot
}
}
