% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlp_sens_analysis.R
\name{SensitivityAnalysis}
\alias{SensitivityAnalysis}
\title{Conduct the sensitivity analysis for each set of parameters.}
\usage{
SensitivityAnalysis(
  clim_dt,
  vegpar_templ,
  vegpar_mat,
  sitepar_templ,
  sitepar_mat,
  model = "pnet-day",
  target_pars = c("GrsPsnMo", "NetPsnMo"),
  annual = TRUE,
  cl = NULL
)
}
\arguments{
\item{clim_dt}{A table that contains monthly climate data.}

\item{vegpar_mat}{Sampled parameter matrix.}

\item{model}{Model name, should be one of "pnet-day", "pnet-ii",
and "pnet-cn".}

\item{target_pars}{Interested target output parameters.}

\item{annual}{Logical. Whether the target parameters are annual.}

\item{vegpar_template}{A template object for the \code{vegpar} object.}

\item{sitepar}{A table that contains site-specific variables.}
}
\value{
Model outputs for the target parameters.
}
\description{
Conduct the sensitivity analysis for each set of parameters.
}
